reformatting suggested at PR review
